move skills to util/components
