+++ dungeon/src/contrib/entities/HeroFactory.java
@@ -9,9 +9,11 @@ import contrib.configuration.KeyboardConfig;
 import contrib.hud.DialogUtils;
 import contrib.hud.elements.GUICombination;
 import contrib.hud.inventory.InventoryGUI;
-import contrib.skill.*;
-import contrib.skill.projectileSkill.FireballSkill;
-import contrib.skill.projectileSkill.TPBallSkill;
+import contrib.utils.components.skill.Resource;
+import contrib.utils.components.skill.Skill;
+import contrib.utils.components.skill.SkillTools;
+import contrib.utils.components.skill.projectileSkill.FireballSkill;
+import contrib.utils.components.skill.projectileSkill.TPBallSkill;
 import contrib.utils.components.health.Damage;
 import contrib.utils.components.interaction.InteractionTool;
 import core.Entity;
