+++ dungeon/src/contrib/item/concreteItem/ItemWoodenBow.java
@@ -4,8 +4,8 @@ import contrib.components.InventoryComponent;
 import contrib.configuration.KeyboardConfig;
 import contrib.entities.WorldItemBuilder;
 import contrib.item.Item;
-import contrib.skill.SkillTools;
-import contrib.skill.projectileSkill.BowSkill;
+import contrib.utils.components.skill.SkillTools;
+import contrib.utils.components.skill.projectileSkill.BowSkill;
 import core.Entity;
 import core.Game;
 import core.components.InputComponent;
