+++ dungeon/src/contrib/utils/components/ai/fight/AIRangeBehaviour.java
@@ -3,7 +3,7 @@ package contrib.utils.components.ai.fight;
 import static core.level.utils.LevelUtils.accessibleTilesInRange;
 
 import com.badlogic.gdx.ai.pfa.GraphPath;
-import contrib.skill.Skill;
+import contrib.utils.components.skill.Skill;
 import contrib.utils.components.ai.AIUtils;
 import contrib.utils.components.ai.ISkillUser;
 import core.Entity;
