+++ dungeon/src/contrib/utils/components/skill/projectileSkill/ProjectileSkill.java
@@ -1,11 +1,11 @@
-package contrib.skill.projectileSkill;
+package contrib.utils.components.skill.projectileSkill;
 
 import contrib.components.CollideComponent;
 import contrib.components.FlyComponent;
 import contrib.components.ProjectileComponent;
-import contrib.skill.Resource;
-import contrib.skill.Skill;
-import contrib.skill.SkillTools;
+import contrib.utils.components.skill.Resource;
+import contrib.utils.components.skill.Skill;
+import contrib.utils.components.skill.SkillTools;
 import core.Entity;
 import core.Game;
 import core.components.DrawComponent;
