+++ dungeon/src/contrib/utils/components/skill/projectileSkill/TPBallSkill.java
@@ -1,4 +1,4 @@
-package contrib.skill.projectileSkill;
+package contrib.utils.components.skill.projectileSkill;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.audio.Sound;
